requirements.txt
openai
python-dotenv
qdrant-client
llama-index
llama-index-vector-stores-qdrant
llama-index-embeddings-ollama
llama-index-embeddings-openai
llama-index-llms-anthropic
llama-index-llms-ollama
llama-index-llms-openai
llama-index-agent-openai
streamlit
langfuse